Fix #71 Assigning each connection to a room in PONG IA
