+++ backend/requirements.txt
@@ -10,7 +10,7 @@ constantly==23.10.4
 cryptography==42.0.7
 daphne==4.1.2
 Django==5.0.6
-django-allauth==0.63.1
+django-allauth==0.62.1
 django-cors-headers==4.3.1
 djangorestframework==3.15.1
 gunicorn==22.0.0
@@ -36,4 +36,4 @@ Twisted==24.3.0
 txaio==23.1.1
 typing_extensions==4.11.0
 urllib3==2.2.1
-zope.interface==6.4
+zope.interface==6.3
